scripts/requirements.txt
amazon-textract-prettyprinter==0.0.10
amazon-textract-response-parser==0.1.20
boto3==1.19.10
botocore==1.22.10
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
jmespath==0.10.0
lxml==4.6.4
marshmallow==3.11.1
Pillow==8.4.0
pymongo==3.12.1
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
tabulate==0.8.9
trp==0.1.6a1
urllib3==1.26.7
